Add files via upload
Attempt at deploying the application. 
Included the header files and Unix shared libraries for Box2D and SFML. Qt Creator finds the header files in the project directories and not in the default include path
Added the "Plugins" directory to the library search path in the application. The app should be able to find shared libraries in this directory if they're not installed in the default locations (to be tested)